feat(gh-repo-settings): ensure all actions are allowed
limiting them is kind of pointless as the main goal of limiting them would be secret prevention. Pull Requests cant steal secrets until they are merged so its kind of pointless.
I already use not that many actions and consider whether to use them in the first place, no need to add the additional complexity here.